Fix build error

+++ src/plugins/hetero/src/subgraph_collector.cpp
@@ -650,7 +650,7 @@ void ov::hetero::fix_model_with_paged_attention(std::shared_ptr<ov::Model>& mode
     }
 
     for (auto& node : vector_visited_transpose) {
-        std::map<int, NodePtr> org_users;
+        std::map<size_t, NodePtr> org_users;
         auto output_shape = node->get_output_partial_shape(0);
         int num_key_value_heads = static_cast<int>(output_shape[2].get_length());
         int head_dim = static_cast<int>(output_shape[3].get_length());
